+++ neurolang/frontend/datalog/natural_syntax.py
@@ -6,7 +6,7 @@ from ...probabilistic.expressions import ProbabilisticPredicate
 from ...datalog import Implication
 from ...datalog.aggregation import AggregationApplication
 from ...expressions import Expression, FunctionApplication
-from .standard_syntax_datalog import (
+from .standard_syntax import (
     DatalogSemantics as DatalogClassicSemantics
 )
