+++ neurolang/frontend/datalog/tests/test_parser_nsd.py
@@ -3,8 +3,8 @@ from operator import add, eq, mul, pow, sub, truediv
 from ....datalog import Conjunction, Fact, Implication, Negation, Union
 from ....datalog.aggregation import AggregationApplication
 from ....expressions import Constant, Symbol
-from ..standard_syntax_datalog import ExternalSymbol
-from ..natural_syntax_datalog import parser
+from ..standard_syntax import ExternalSymbol
+from ..natural_syntax import parser
 from ....probabilistic.expressions import ProbabilisticPredicate
