+++ neurolang/frontend/query_resolution_datalog.py
@@ -26,8 +26,8 @@ from ..datalog.expression_processing import (
 )
 from ..type_system import Unknown
 from ..utils import NamedRelationalAlgebraFrozenSet, RelationalAlgebraFrozenSet
-from .datalog.standard_syntax_datalog import parser as datalog_parser
-from .datalog.natural_syntax_datalog import parser as nat_datalog_parser
+from .datalog.standard_syntax import parser as datalog_parser
+from .datalog.natural_syntax import parser as nat_datalog_parser
 from .query_resolution import NeuroSynthMixin, QueryBuilderBase, RegionMixin
 from ..datalog import DatalogProgram
 from . import query_resolution_expressions as fe
